azure_requirements.txt
# Azure dependencies for storage account retrieval
azure-identity>=1.12.0
azure-mgmt-storage>=21.0.0